Refactor reducer factories into one rootReducer
