+++ src/reducers/filter.js
@@ -1,13 +1,15 @@
-const filterReducerFactory = initialState => (state = initialState, action) => {
-  switch(action.type) {
+const filterReducer = (state, action) => {
+  switch (action.type) {
     case 'CHANGE_FILTER': {
       return {
         ...state,
         filter: action.filter,
-      }
+      };
     }
 
     default:
       return state;
   }
-}
+};
+
+export default filterReducer;
